remove wrap-root-element file, put contents in gatsby-ssr
